add type Decimal for calculations

+++ packages/hardhat/contracts/PositionTypes.sol
@@ -1,15 +1,17 @@
 // SPDX-License-Identifier: MIT
 pragma solidity >=0.8.0 <0.9.0;
 
+import "./DecimalMath.sol";
+
 library PositionTypes {
     struct Position {
         uint256 inBalance;
         uint256 shares;
-        uint256 index;
+        Decimal index;
         uint256 lastUpdateTime;
-        uint256 pendingReward;
+        Decimal pendingReward;
         uint256 spentIn;
         uint256 purchased;
         uint256 exitDate;
     }
-} 
+}
